destructure-methods {zeallot}	R Documentation

Included Implementations of destructure

Description

zeallot includes destructure methods for the following classes: character, complex, Date, data.frame, and summary.lm. See details for how each object is transformed into a list.

Usage

## S3 method for class 'character'
destructure(x)

## S3 method for class 'complex'
destructure(x)

## S3 method for class 'Date'
destructure(x)

## S3 method for class 'data.frame'
destructure(x)

## S3 method for class 'summary.lm'
destructure(x)

## Default S3 method:
destructure(x)

Arguments

x
An R object.

Details

character values are split into a list of individual characters.

complex values are split into a list of two values, the real and the imaginary part.

Date values are split into a list of three numeric values, the year, month, and day.

data.frame values are coerced into a list using as.list.

summary.lm values are coerced into a list of values, one element for each of the eleven values returned by summary.lm.

Value

A list of elements from x.

See Also

destructure

[Package zeallot version 0.1.0 Index]