zeallot {zeallot}	R Documentation

Multiple, unpacking, and destructuring assignment in R

Description

zeallot provides a %<-% operator to perform multiple assignment in R. To get started with zeallot be sure to read over the introductory vignette on unpacking assignment, vignette('unpacking-assignment').

Author(s)

Maintainer: Nathan Teetor nate@haufin.ch

Other contributors:

Paul Teetor [contributor]

See Also

%<-%

[Package zeallot version 0.2.0 Index]