zeallous {zeallot}	R Documentation

Allow zeallous assignment

Description

Using zeallot within an R package may cause R CMD check to raise NOTEs concerning variables with "no visible binding". To avoid these NOTEs, include a call to zeallous() in a package's .onLoad() function.

Usage

zeallous()

Details

The R CMD check process uses a package {codetools} to check for assigned variables. However, due to the non-standard nature of zeallot assignment the codetools package does not identify these variables. To work around this, the zeallous() function modifies the variables found by codetools to avoid the NOTEs raised by R CMD check.

Examples

.onLoad <- function(libname, pkgname) {
  zeallous()
}

[Package zeallot version 0.2.0 Index]